Scratch Addons Compatibility Api Function
I added an API function for use in features that require checking if a conflicting Scratch Addons addon is enabled. If a specified addon is enabled, it will return `true` and `undefined` if not (or if Scratch Addons aren't enabled/installed in the browser at all). Someone creating or fixing a Scratch Tools feature that doesn't work well with another function or functions from Scratch Addons can use this to fix the problem by customizing feature behavior.