Outdated feature.json on their branch
They has an outdated features.json file on their branch, user didn't seem to be active, so I didn't bother making a pr on their branch to update it. Thought it would be easier to update it here, also would make a good v3.4.0 feature so I thought i'd finish it.