Update wording for clarity
I tested it and the feature works fine in displaying comments that are the full 500 characters (scratch's project comment limit). The original wording, "view full comment" can then be considered confusing to the user, as they might think that the full comment isn't being displayed/is being cut off when it's just the context (replies + original message) that is missing.